gh-108303: Move `ann_module*.py` files to `typinganndata/` folder (#108354)
(cherry picked from commit 3f61cf646d0506baa0c0c2118f05110446519c62)
