+++ Lib/test/test_opcodes.py
@@ -1,7 +1,8 @@
 # Python test set -- part 2, opcodes
 
 import unittest
-from test import ann_module, support
+from test import support
+from test.typinganndata import ann_module
 
 class OpcodeTest(unittest.TestCase):
